custom_components/mazda/requirements.txt
aiohttp==3.10.5
cryptography==43.0.1
ha-mqtt-discoverable==0.14.0
paho-mqtt==1.6.1
python-dotenv==1.0.1
